refactor: SearchButtonTriggerSpotlight로 통합
